non_hierarch: proliferate change in types for form,lev
Restores a batch of SBO tests

+++ src/NonDExpansion.hpp
@@ -255,7 +255,7 @@ protected:
   void multilevel_regression();
 
   /// configure fidelity/level counts from model hierarchy
-  void configure_sequence(size_t& num_steps, unsigned short& fixed_index,
+  void configure_sequence(size_t& num_steps, size_t& fixed_index,
 			  short& seq_type);
   /// extract cost estimates from model hierarchy (forms or resolutions)
   void configure_cost(unsigned short num_steps, bool multilevel,
@@ -263,10 +263,10 @@ protected:
   /// extract cost estimates from model hierarchy, if available
   bool query_cost(unsigned short num_steps, bool multilevel, RealVector& cost);
   /// configure response mode and active/truth/surrogate model keys within a
-  /// hierarchical model.  s_index is the sequence index that defines the
-  /// active dimension for a model sequence.
+  /// hierarchical model.  seq_type is the type of sequence that defines the
+  /// active dimension for traversing a model sequence.
   void configure_indices(unsigned short group, unsigned short form,
-			 unsigned short lev,   short seq_type);
+			 size_t lev,           short seq_type);
   /// return aggregate cost (one or more models) for a level sample
   Real sequence_cost(unsigned short step, const RealVector& cost);
   /// compute equivHFEvals from samples per level and cost per evaluation
